requirements.txt
requests
mysql-connector-python
pytorch
torchvision